Revert "Duration and Frequency to float type"
This reverts commit 88eb49ff52d28d71b298bae3f838baaf020b09cd.

+++ R2C2-RepRap_firmware/libraries/R2C2/buzzer.h
@@ -37,10 +37,10 @@ void buzzer_init (void);
  * 20Hz <= frequency <= 20000
  * 10ms <= duration <= (2^16 - 1)ms
  */
-void buzzer_play (float frequency, uint32_t duration);
+void buzzer_play (uint16_t frequency, uint16_t duration);
 void buzzer_wait (void);
 
-void buzzer_play_sync (float frequency, uint32_t duration);
+void buzzer_play_sync (uint16_t frequency, uint32_t duration);
 
 #endif  /* _BUZZER_ */
